Allow slurm_opt_get_tres_per_task_cpu_cnt() to parse "cpu=" and "cpu:"

Closes #50176

See merge request SchedMD/dev/slurm!325

+++ src/common/slurm_opt.h
@@ -460,7 +460,7 @@ extern bool slurm_option_get_tres_per_tres(
 	char *in_val, char *tres_name, uint64_t *cnt, char **save_ptr, int *rc);
 
 /* Get cpus per task count from tres per task string */
-extern int slurm_opt_get_tres_per_task_cpu_cnt(char *tres_per_task);
+extern uint16_t slurm_opt_get_tres_per_task_cpu_cnt(char *tres_per_task);
 
 extern struct option *slurm_option_table_create(slurm_opt_t *opt,
 						char **opt_string);
